Moved expansion handler to view model
